requirements.txt

Flask==2.3.3
pyyaml==6.0